Fix board and row tests and move row module into board module
