+++ src/board/row/factory/__tests__/getRows.test.ts
@@ -1,63 +1,51 @@
 import { getRow } from "../getRow";
 
 describe("Given the function getRow", () => {
-  describe("When it receives 3", () => {
+  describe("When it receives 3 as number of cells", () => {
     test("Then it should return a row of 3 cells", () => {
       const cellsNumber = 3;
-      const expectedCellsNumber = 3;
 
       const actualRow = getRow(cellsNumber);
 
-      expect(expectedCellsNumber).toBe(actualRow.length);
+      expect(actualRow.length).toBe(cellsNumber);
     });
   });
 
-  describe("When it receives 4", () => {
+  describe("When it receives 4 as number of cells", () => {
     test("Then it should return a row of 6 cells", () => {
       const cellsNumber = 4;
-      const expectedCellsNumber = 4;
 
       const actualRow = getRow(cellsNumber);
 
-      expect(expectedCellsNumber).toBe(actualRow.length);
+      expect(actualRow.length).toBe(cellsNumber);
     });
   });
 
-  describe("When it receives 5.9", () => {
+  describe("When it receives 5.9 as number of cells", () => {
     test("Then it should return a row of 6 cells", () => {
       const cellsNumber = 5.9;
       const expectedCellsNumber = 6;
 
       const actualRow = getRow(cellsNumber);
 
-      expect(expectedCellsNumber).toBe(actualRow.length);
+      expect(actualRow.length).toBe(expectedCellsNumber);
     });
   });
 
-  describe("When it receives 6.1", () => {
+  describe("When it receives 6.1 as number of cells", () => {
     test("Then it should return a row of 6 cells", () => {
       const cellsNumber = 6.1;
       const expectedCellsNumber = 6;
 
       const actualRow = getRow(cellsNumber);
 
-      expect(expectedCellsNumber).toBe(actualRow.length);
+      expect(actualRow.length).toBe(expectedCellsNumber);
     });
   });
 
-  describe("When it receives 2", () => {
+  describe("When it receives 2 as number of cells", () => {
     test("Then it should throw an error", () => {
       const cellsNumber = 2;
-
-      expect(() => {
-        getRow(cellsNumber);
-      }).toThrow();
-    });
-  });
-
-  describe("When it receives 0", () => {
-    test("Then it should throw an error", () => {
-      const cellsNumber = 0;
 
       expect(() => {
         getRow(cellsNumber);
