+++ src/board/row/types.ts
@@ -1,0 +1,3 @@
+import { Cell } from "../../cell/types";
+
+export type Row = Cell[];
