+++ src/board/types.ts
@@ -1,3 +1,3 @@
-import { Row } from "../row/types";
+import { Row } from "./row/types";
 
 export type Board = Row[];
