--- src/row/types.ts
@@ -1,3 +1,0 @@
-import { Cell } from "../cell/types";
-
-export type Row = Cell[];
